requirements.txt
streamlit
pdf2docx
Pillow
pandas
fpdf
docx
PyPDF2
pdf2image
openpyxl
